Move LambdaMetadata to separate file
